Добавлены новые прокачки в лабораторию

+++ node/src/main/kotlin/me/reidj/tower/laboratory/Research.kt
@@ -4,7 +4,7 @@ package me.reidj.tower.laboratory
  * @project tower
  * @author Рейдж
  */
-class Research(private val researchType: ResearchType, var level: Int, var whenBought: Int?) {
+class Research(private val researchType: ResearchType, var level: Int, var whenBought: Double?) {
 
     fun getValue() = researchType.value + researchType.step * level
 }
